constructor

+++ src/main/java/com/application_server/server/restController/AdminRestController.java
@@ -30,7 +30,7 @@ public class AdminRestController {
     public ResponseEntity<List<UserDto>> showAllUser() {
         List<UserDto> listUserDto = userService.findAll().stream()
                 .map(userFactory::create).collect(Collectors.toList());
-        if (listUserDto.isEmpty()){
+        if (listUserDto.isEmpty()) {
             return new ResponseEntity<>(HttpStatus.NOT_FOUND);
         }
         return new ResponseEntity<>(listUserDto, HttpStatus.OK);
@@ -38,23 +38,18 @@ public class AdminRestController {
 
 
     @PostMapping("/create")
-    public ResponseEntity<UserDto> createUser(@RequestBody UserDto user){
+    public ResponseEntity<UserDto> createUser(@RequestBody UserDto user) {
         if (user == null) {
-            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
-        }
-        User newUser = userService.addUser(user);
-
-        if (newUser == null){
             return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
         }
-
+        User newUser = userService.addUser(user);
         return new ResponseEntity<>(userFactory.create(newUser), HttpStatus.CREATED);
     }
 
     @PutMapping("/edit")
     public ResponseEntity<UserDto> editUser(@RequestBody UserDto user) {
         if (user == null) {
-            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+            return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
         }
         Long id = user.getId();
         User updateUser = userService.updateUserById(id, user);
@@ -64,7 +59,7 @@ public class AdminRestController {
     @GetMapping("/{id}")
     public ResponseEntity<UserDto> getUser(@PathVariable Long id) {
         if (userService.findUserById(id) == null) {
-            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+            return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
         }
         User user = userService.findUserById(id);
         return new ResponseEntity<>(userFactory.create(user), HttpStatus.OK);
@@ -72,8 +67,12 @@ public class AdminRestController {
 
     @DeleteMapping("/{id}")
     public ResponseEntity<String> deleteUser(@PathVariable Long id) {
+        if (id == null) {
+            return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+        }
         userService.deleteUserById(id);
         return new ResponseEntity<>("User was deleted ...", HttpStatus.NO_CONTENT);
     }
+
 }
